@@ -5,7 +5,7 @@ class Program
 {
     static void Main()
     {
-        var p = new Program();
+        Program p = new();
         p.Log();
         p.SomeProperty = 33;
         Action a1 = () => p.Log();
@@ -24,12 +24,12 @@ class Program
     }
 
     public void Log([CallerLineNumber] int line = -1,
-        [CallerFilePath] string? path = null,
-        [CallerMemberName] string? name = null)
+        [CallerFilePath] string? path = default,
+        [CallerMemberName] string? name = default)
     {
-        Console.WriteLine((line < 0) ? "No line" : "Line " + line);
-        Console.WriteLine((path == null) ? "No file path" : path);
-        Console.WriteLine((name == null) ? "No member name" : name);
+        Console.WriteLine($"Line {line}");
+        Console.WriteLine(path);
+        Console.WriteLine(name);
         Console.WriteLine();
     }
 }
