@@ -3,12 +3,8 @@
 public class ColdCallFileFormatException : Exception
 {
     public ColdCallFileFormatException(string message)
-        : base(message)
-    {
-    }
+        : base(message) {}
 
     public ColdCallFileFormatException(string message, Exception innerException)
-        : base(message, innerException)
-    {
-    }
+        : base(message, innerException) {}
 }
