@@ -2,7 +2,6 @@
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using System;
 
 using var host = Host.CreateDefaultBuilder(args)
     .ConfigureServices((context, services) =>
@@ -19,19 +18,11 @@ using var host = Host.CreateDefaultBuilder(args)
 using var scope = host.Services.CreateScope();
 var runner = scope.ServiceProvider.GetRequiredService<Runner>();
 
-string input = Console.ReadLine()?.Trim().ToLower() ?? string.Empty;
-if (input == "yes" || input == "y")
-{
-    await runner.ApplyMigrationsAsync();
-}
-else
-{
-    await runner.CreateTheDatabaseAsync();
-}
-
-await runner.AddBookAsync("sample book", "sample");
+await runner.CreateTheDatabaseAsync();
+await runner.AddBookAsync("Professional C# and .NET", "Wrox Press");
 await runner.AddBooksAsync();
 await runner.ReadBooksAsync();
 await runner.QueryBooksAsync();
+await runner.UpdateBookAsync();
 await runner.DeleteBooksAsync();
 await runner.DeleteDatabaseAsync();
