@@ -32,27 +32,22 @@ public class Runner
         }
     }
 
-    public async Task ApplyMigrationsAsync()
-    {
-        await _booksContext.Database.MigrateAsync();
-    }
-
     public async Task AddBookAsync(string title, string publisher)
     {
         Book book = new(title, publisher);
         await _booksContext.Books.AddAsync(book);
         int records = await _booksContext.SaveChangesAsync();
-        Console.WriteLine($"{records} record added");
+        Console.WriteLine($"{records} record added with id {book.BookId}");
 
         Console.WriteLine();
     }
 
     public async Task AddBooksAsync()
     {
-        Book b1 = new("Professional C# 6 and .NET Core 1.0", "Wrox Press");
-        Book b2 = new("Professional C# 5 and .NET 4.5.1", "Wrox Press");
-        Book b3 = new("JavaScript for Kids", "Wrox Press");
-        Book b4 = new("Web Design with HTML and CSS", "For Dummies");
+        Book b1 = new("Professional C# 7 and .NET Core 2", "Wrox Press");
+        Book b2 = new("Professional C# 6 and .NET Core 1.0", "Wrox Press");
+        Book b3 = new("Professional C# 5 and .NET 4.5.1", "Wrox Press");
+        Book b4 = new("Essential Algorithms", "Wiley");
         await _booksContext.Books.AddRangeAsync(b1, b2, b3, b4);
         int records = await _booksContext.SaveChangesAsync();
         Console.WriteLine($"{records} records added");
@@ -60,9 +55,13 @@ public class Runner
         Console.WriteLine();
     }
 
-    public async Task ReadBooksAsync(CancellationToken cancel = default)
+    public async Task ReadBooksAsync(CancellationToken token = default)
     {
-        List<Book> books = await _booksContext.Books.ToListAsync(cancel);
+#if DEBUG
+        string query = _booksContext.Books.ToQueryString();
+        Console.WriteLine(query);
+#endif
+        List<Book> books = await _booksContext.Books.ToListAsync(token);
         foreach (var b in books)
         {
             Console.WriteLine($"{b.Title} {b.Publisher}");
@@ -73,6 +72,8 @@ public class Runner
 
     public async Task QueryBooksAsync(CancellationToken token = default)
     {
+        string query = _booksContext.Books.Where(b => b.Publisher == "Wrox Press").ToQueryString();
+        Console.WriteLine(query);
         await _booksContext.Books
             .Where(b => b.Publisher == "Wrox Press")
             .ForEachAsync(b =>
@@ -83,21 +84,37 @@ public class Runner
         Console.WriteLine();
     }
 
+#if USERECORDS
     public async Task UpdateBookAsync()
     {
-        int records = 0;
-        Book book = await _booksContext.Books
-            .Where(b => b.Title == "Professional C# 7")
-            .FirstOrDefaultAsync();
+        Book? book = await _booksContext.Books.FindAsync(1);
+
         if (book != null)
         {
-            Book bookUpdate = book with { Title = "Professional C# 7 and .NET Core 2" };
+            // detach the existing object from the context which allows to attach it with the Update method
+            _booksContext.Entry(book).State = EntityState.Detached;
+            Book bookUpdate = book with { Title = "Professional C# and .NET - 2021 Edition" };
             _booksContext.Update(bookUpdate);
-            records = await _booksContext.SaveChangesAsync();
+            int records = await _booksContext.SaveChangesAsync();
+            Console.WriteLine($"{records} record updated");
+
         }
-        Console.WriteLine($"{records} record updated");
         Console.WriteLine();
     }
+#else
+    public async Task UpdateBookAsync()
+    {
+        Book? book = _booksContext.Books.Find(1);
+
+        if (book != null)
+        {
+            book.Title = "Professional C# and .NET - 2021 Edition";
+            int records = await _booksContext.SaveChangesAsync();
+            Console.WriteLine($"{records} record updated");
+        }
+        Console.WriteLine();
+    }
+#endif
 
     public async Task DeleteBooksAsync()
     {
